+++ lib/MIP/Cli/Mip/Analyse/Rd_dna.pm
@@ -1157,6 +1157,115 @@ q{Default: hgvs, symbol, numbers, sift, polyphen, humdiv, domains, protein, ccds
     );
 
     option(
+        q{me_merge_bam} => (
+            cmd_tags      => [q{Analysis recipe switch}],
+            documentation => q{Prepare bam files for RetroSeq},
+            is            => q{rw},
+            isa           => enum( [ 0, 1, 2 ] ),
+        )
+    );
+
+    option(
+        q{retroseq} => (
+            cmd_tags      => [q{Analysis recipe switch}],
+            documentation => q{Discover mobile elements using RetroSeq},
+            is            => q{rw},
+            isa           => enum( [ 0, 1, 2 ] ),
+        )
+    );
+
+    option(
+        q{me_merge_vcfs} => (
+            cmd_tags      => [q{Analysis recipe switch}],
+            documentation => q{Merge sample vcfs from RetroSeq},
+            is            => q{rw},
+            isa           => enum( [ 0, 1, 2 ] ),
+        )
+    );
+
+    option(
+        q{me_merge_vcfs_bnd_distance} => (
+            cmd_tags      => [q{Default: 150}],
+            documentation => q{Maximum distance between two similar BNDs},
+            is            => q{rw},
+            isa           => Num,
+        )
+    );
+
+    option(
+        q{me_merge_vcfs_overlap} => (
+            cmd_tags      => [q{Default: }],
+            documentation => q{Overlap required to merge two events},
+            is            => q{rw},
+            isa           => Num,
+        )
+    );
+
+    option(
+        q{mobile_element_reference} => (
+            cmd_tags      => [q{file.vcf=TE_type}],
+            documentation => q{Database file(s) for mobile element iscovery},
+            is            => q{rw},
+            isa           => HashRef,
+        )
+    );
+
+    option(
+        q{me_annotate} => (
+            cmd_tags      => [q{Analysis recipe switch}],
+            documentation => q{Annotate mobile elememnt},
+            is            => q{rw},
+            isa           => enum( [ 0, 1, 2 ] ),
+        )
+    );
+
+    option(
+        q{me_annotate_query_bnd_distance} => (
+            cmd_tags      => [q{Default: 150}],
+            documentation => q{Maximum distance between two similar BNDs},
+            is            => q{rw},
+            isa           => Num,
+        )
+    );
+
+    option(
+        q{me_annotate_query_overlap} => (
+            cmd_tags      => [q{Default: }],
+            documentation => q{Overlap required to annotate},
+            is            => q{rw},
+            isa           => Num,
+        )
+    );
+
+    option(
+        q{me_varianteffectpredictor} => (
+            cmd_tags      => [q{Analysis recipe switch}],
+            documentation => q{Annotate mobile elements with VEP},
+            is            => q{rw},
+            isa           => enum( [ 0, 1, 2 ] ),
+        )
+    );
+
+    option(
+        q{me_filter} => (
+            cmd_tags      => [q{Analysis recipe switch}],
+            documentation => q{Filter mobile elements},
+            is            => q{rw},
+            isa           => enum( [ 0, 1, 2 ] ),
+        )
+    );
+
+    option(
+        q{me_filter_frequency_threshold} => (
+            cmd_tags      => [q{Default: 0.1}],
+            documentation =>
+              q{Threshold frequency for variants to be filtered out, set to 0 to disable},
+            is  => q{rw},
+            isa => Num,
+        )
+    );
+
+    option(
         q{gatk_haplotypecaller} => (
             cmd_tags      => [q{Analysis recipe switch}],
             documentation => q{Variant discovery using GATK HaplotypeCaller},