+++ lib/MIP/Program/Retroseq.pm
@@ -36,7 +36,7 @@ sub retroseq_call {
 ## Returns  : @commands
 ## Arguments: $filehandle              => Filehandle to write to
 ##          : $infile_path             => Path to input bam file
-##          : $outputfile_path         => Path to the output file
+##          : $outfile_path            => Path to the output file
 ##          : $reference_fasta_path    => Reference genome path
 ##          : $retroseq_bed_path       => Path to the retroseq discover bedfile
 ##          : $stderrfile_path         => Stderrfile path
@@ -48,7 +48,7 @@ sub retroseq_call {
     ## Flatten argument(s)
     my $filehandle;
     my $infile_path;
-    my $outputfile_path;
+    my $outfile_path;
     my $reference_fasta_path;
     my $retroseq_bed_path;
     my $stderrfile_path;
@@ -65,10 +65,10 @@ sub retroseq_call {
             store       => \$infile_path,
             strict_type => 1,
         },
-        outputfile_path => {
+        outfile_path => {
             defined     => 1,
             required    => 1,
-            store       => \$outputfile_path,
+            store       => \$outfile_path,
             strict_type => 1,
         },
         reference_fasta_path => {
@@ -109,7 +109,7 @@ sub retroseq_call {
 
     push @commands, q{-ref} . $SPACE . $reference_fasta_path;
 
-    push @commands, q{-output} . $SPACE . $outputfile_path;
+    push @commands, q{-output} . $SPACE . $outfile_path;
 
     push @commands,
       unix_standard_streams(
@@ -138,7 +138,7 @@ sub retroseq_discover {
 ## Arguments: $filehandle              => Filehandle to write to
 ##          : $infile_path             => Path to input bam file
 ##          : $mobile_element_tsv_path => Tab separated file containing the name and path of mobile elements
-##          : $outputfile_path         => path to the output file
+##          : $outfile_path            => path to the output file
 ##          : $stderrfile_path         => Stderrfile path
 ##          : $stderrfile_path_append  => Append stderr info to file path
 ##          : $stdoutfile_path         => Stdoutfile path
@@ -149,7 +149,7 @@ sub retroseq_discover {
     my $filehandle;
     my $infile_path;
     my $mobile_element_tsv_path;
-    my $outputfile_path;
+    my $outfile_path;
     my $stderrfile_path;
     my $stderrfile_path_append;
     my $stdoutfile_path;
@@ -170,10 +170,10 @@ sub retroseq_discover {
             store       => \$mobile_element_tsv_path,
             strict_type => 1,
         },
-        outputfile_path => {
+        outfile_path => {
             defined     => 1,
             required    => 1,
-            store       => \$outputfile_path,
+            store       => \$outfile_path,
             strict_type => 1,
         },
         stderrfile_path => {
@@ -200,7 +200,7 @@ sub retroseq_discover {
 
     push @commands, q{-refTEs} . $SPACE . $mobile_element_tsv_path;
 
-    push @commands, q{-output} . $SPACE . $outputfile_path;
+    push @commands, q{-output} . $SPACE . $outfile_path;
 
     push @commands,
       unix_standard_streams(