+++ t/analysis_me_annotate.t
@@ -16,11 +16,12 @@ use warnings qw{ FATAL utf8 };
 ## CPANM
 use autodie qw { :all };
 use Modern::Perl qw{ 2018 };
+use Readonly;
 use Test::Trap;
 
 ## MIPs lib/
 use lib catdir( dirname($Bin), q{lib} );
-use MIP::Constants qw{ $COLON $COMMA $SPACE set_container_constants };
+use MIP::Constants qw{ $COLON $COMMA $SPACE };
 use MIP::Test::Fixtures qw{ test_add_io_for_recipe test_log test_mip_hashes };
 
 BEGIN {
@@ -30,16 +31,16 @@ BEGIN {
 ### Check all internal dependency modules and imports
 ## Modules with import
     my %perl_module = (
-        q{MIP::Recipes::Analysis::Deepvariant} => [qw{ analysis_deepvariant }],
+        q{MIP::Recipes::Analysis::Me_annotate} => [qw{ analysis_me_annotate }],
         q{MIP::Test::Fixtures} => [qw{ test_add_io_for_recipe test_log test_mip_hashes }],
     );
 
     test_import( { perl_module_href => \%perl_module, } );
 }
 
-use MIP::Recipes::Analysis::Deepvariant qw{ analysis_deepvariant };
+use MIP::Recipes::Analysis::Me_annotate qw{ analysis_me_annotate };
 
-diag(   q{Test analysis_deepvariant from Deepvariant.pm}
+diag(   q{Test analysis_me_annotate from Me_annotate.pm}
       . $COMMA
       . $SPACE . q{Perl}
       . $SPACE
@@ -47,10 +48,14 @@ diag(   q{Test analysis_deepvariant from Deepvariant.pm}
       . $SPACE
       . $EXECUTABLE_NAME );
 
-test_log( { log_name => q{MIP}, no_screen => 1, } );
+## Constants
+Readonly my $ANNOTATION_OVERLAP => -1;
+Readonly my $BND_DISTANCE       => 150;
+
+test_log( { no_screen => 1, } );
 
 ## Given analysis parameters
-my $recipe_name    = q{deepvariant};
+my $recipe_name    = q{me_annotate};
 my $slurm_mock_cmd = catfile( $Bin, qw{ data modules slurm-mock.pl } );
 
 my %active_parameter = test_mip_hashes(
@@ -59,19 +64,13 @@ my %active_parameter = test_mip_hashes(
         recipe_name   => $recipe_name,
     }
 );
-
 $active_parameter{$recipe_name}                     = 1;
 $active_parameter{recipe_core_number}{$recipe_name} = 1;
-$active_parameter{recipe_gpu_number}{$recipe_name}  = 1;
 $active_parameter{recipe_time}{$recipe_name}        = 1;
-$active_parameter{container_manager}                = q{singularity};
-my $sample_id = $active_parameter{sample_ids}[0];
-
-set_container_constants(
-    {
-        active_parameter_href => \%active_parameter,
-    }
-);
+$active_parameter{me_annotate_query_files}          = { a_file => q{a_file|AF|AC|in_AF|in_AC|1}, };
+my $case_id = $active_parameter{case_id};
+$active_parameter{me_annotate_query_overlap}      = $ANNOTATION_OVERLAP;
+$active_parameter{me_annotate_query_bnd_distance} = $BND_DISTANCE;
 
 my %file_info = test_mip_hashes(
     {
@@ -91,28 +90,32 @@ my %parameter = test_mip_hashes(
 test_add_io_for_recipe(
     {
         file_info_href => \%file_info,
-        id             => $sample_id,
+        id             => $case_id,
         parameter_href => \%parameter,
         recipe_name    => $recipe_name,
-        step           => q{bam},
+        step           => q{vcf},
     }
 );
 
-my %sample_info;
+my %sample_info = test_mip_hashes(
+    {
+        mip_hash_name => q{qc_sample_info},
+        recipe_name   => $recipe_name,
+    }
+);
 
-my $is_ok = analysis_deepvariant(
+my $is_ok = analysis_me_annotate(
     {
         active_parameter_href => \%active_parameter,
+        case_id               => $case_id,
         file_info_href        => \%file_info,
         job_id_href           => \%job_id,
         parameter_href        => \%parameter,
         profile_base_command  => $slurm_mock_cmd,
         recipe_name           => $recipe_name,
-        sample_id             => $sample_id,
         sample_info_href      => \%sample_info,
     }
 );
-
 ## Then return TRUE
 ok( $is_ok, q{ Executed analysis recipe } . $recipe_name );