+++ t/analysis_vep_me.t
@@ -16,12 +16,18 @@ use warnings qw{ FATAL utf8 };
 ## CPANM
 use autodie qw { :all };
 use Modern::Perl qw{ 2018 };
+use Readonly;
 use Test::Trap;
 
 ## MIPs lib/
 use lib catdir( dirname($Bin), q{lib} );
-use MIP::Constants qw{ $COLON $COMMA $SPACE set_container_constants };
+use MIP::Constants qw{ $COLON $COMMA $SPACE };
 use MIP::Test::Fixtures qw{ test_add_io_for_recipe test_log test_mip_hashes };
+
+## Constants
+Readonly my $GENOME_BUILD_VERSION_38 => 38;
+Readonly my $GENOME_BUILD_VERSION_20 => 20;
+Readonly my $GENOME_BUILD_VERSION_19 => 19;
 
 BEGIN {
 
@@ -30,16 +36,16 @@ BEGIN {
 ### Check all internal dependency modules and imports
 ## Modules with import
     my %perl_module = (
-        q{MIP::Recipes::Analysis::Deepvariant} => [qw{ analysis_deepvariant }],
-        q{MIP::Test::Fixtures} => [qw{ test_add_io_for_recipe test_log test_mip_hashes }],
+        q{MIP::Recipes::Analysis::Vep} => [qw{ analysis_vep_me }],
+        q{MIP::Test::Fixtures}         => [qw{ test_add_io_for_recipe test_log test_mip_hashes }],
     );
 
     test_import( { perl_module_href => \%perl_module, } );
 }
 
-use MIP::Recipes::Analysis::Deepvariant qw{ analysis_deepvariant };
+use MIP::Recipes::Analysis::Vep qw{ analysis_vep_me };
 
-diag(   q{Test analysis_deepvariant from Deepvariant.pm}
+diag(   q{Test analysis_vep_me from Vep.pm}
       . $COMMA
       . $SPACE . q{Perl}
       . $SPACE
@@ -47,10 +53,10 @@ diag(   q{Test analysis_deepvariant from Deepvariant.pm}
       . $SPACE
       . $EXECUTABLE_NAME );
 
-test_log( { log_name => q{MIP}, no_screen => 1, } );
+my $log = test_log( { log_name => q{MIP}, no_screen => 1, } );
 
 ## Given analysis parameters
-my $recipe_name    = q{deepvariant};
+my $recipe_name    = q{me_varianteffectpredictor};
 my $slurm_mock_cmd = catfile( $Bin, qw{ data modules slurm-mock.pl } );
 
 my %active_parameter = test_mip_hashes(
@@ -59,26 +65,20 @@ my %active_parameter = test_mip_hashes(
         recipe_name   => $recipe_name,
     }
 );
-
 $active_parameter{$recipe_name}                     = 1;
 $active_parameter{recipe_core_number}{$recipe_name} = 1;
-$active_parameter{recipe_gpu_number}{$recipe_name}  = 1;
 $active_parameter{recipe_time}{$recipe_name}        = 1;
-$active_parameter{container_manager}                = q{singularity};
-my $sample_id = $active_parameter{sample_ids}[0];
-
-set_container_constants(
-    {
-        active_parameter_href => \%active_parameter,
-    }
-);
-
+my $case_id = $active_parameter{case_id};
+$active_parameter{vep_directory_cache} = q{a_cache};
 my %file_info = test_mip_hashes(
     {
         mip_hash_name => q{file_info},
         recipe_name   => $recipe_name,
     }
 );
+
+$file_info{human_genome_reference_source}  = q{grch};
+$file_info{human_genome_reference_version} = $GENOME_BUILD_VERSION_38;
 
 my %job_id;
 my %parameter = test_mip_hashes(
@@ -91,24 +91,25 @@ my %parameter = test_mip_hashes(
 test_add_io_for_recipe(
     {
         file_info_href => \%file_info,
-        id             => $sample_id,
+        id             => $case_id,
         parameter_href => \%parameter,
         recipe_name    => $recipe_name,
-        step           => q{bam},
+        step           => q{vcf},
     }
 );
+$active_parameter{vep_features} = [qw{ refseq }];
 
 my %sample_info;
 
-my $is_ok = analysis_deepvariant(
+my $is_ok = analysis_vep_me(
     {
         active_parameter_href => \%active_parameter,
+        case_id               => $case_id,
         file_info_href        => \%file_info,
         job_id_href           => \%job_id,
         parameter_href        => \%parameter,
         profile_base_command  => $slurm_mock_cmd,
         recipe_name           => $recipe_name,
-        sample_id             => $sample_id,
         sample_info_href      => \%sample_info,
     }
 );