+++ t/retroseq_call.t
@@ -23,16 +23,13 @@ use lib catdir( dirname($Bin), q{lib} );
 use MIP::Constants qw{ $COMMA $SPACE };
 use MIP::Test::Commands qw{ test_function };
 
-
 BEGIN {
 
     use MIP::Test::Fixtures qw{ test_import };
 
 ### Check all internal dependency modules and imports
 ## Modules with import
-    my %perl_module = (
-        q{MIP::Program::Retroseq} => [qw{ retroseq_call }],
-);
+    my %perl_module = ( q{MIP::Program::Retroseq} => [qw{ retroseq_call }], );
 
     test_import( { perl_module_href => \%perl_module, } );
 }
@@ -86,7 +83,7 @@ my %required_argument = (
         input           => q{reference.fa},
         expected_output => q{-ref} . $SPACE . q{reference.fa},
     },
-    outputfile_path => {
+    outfile_path => {
         input           => q{HappyKangaroo.vcf},
         expected_output => q{-output} . $SPACE . q{HappyKangaroo.vcf},
     },
@@ -106,7 +103,7 @@ my %specific_argument = (
         input           => q{reference.fa},
         expected_output => q{-ref} . $SPACE . q{reference.fa},
     },
-    outputfile_path => {
+    outfile_path => {
         input           => q{HappyKangaroo.vcf},
         expected_output => q{-output} . $SPACE . q{HappyKangaroo.vcf},
     },